+++ src/commands/moveToFinished-14.lua
@@ -182,8 +182,9 @@ if rcall("EXISTS", jobIdKey) == 1 then -- // Make sure job exists
 
         local target, isPausedOrMaxed = getTargetQueueList(metaKey, KEYS[2], KEYS[1], KEYS[8])
 
+        local markerKey = KEYS[14]
         -- Check if there are delayed jobs that can be promoted
-        promoteDelayedJobs(KEYS[7], KEYS[14], target, KEYS[3], eventStreamKey, prefix,
+        promoteDelayedJobs(KEYS[7], markerKey, target, KEYS[3], eventStreamKey, prefix,
                            timestamp, KEYS[10], isPausedOrMaxed)
 
         local maxJobs = tonumber(opts['limiter'] and opts['limiter']['max'])
@@ -199,12 +200,12 @@ if rcall("EXISTS", jobIdKey) == 1 then -- // Make sure job exists
 
         if jobId then
             return prepareJobForProcessing(prefix, KEYS[6], eventStreamKey, jobId,
-                                            timestamp, maxJobs, opts)
+                                            timestamp, maxJobs, markerKey, opts)
         else
             jobId = moveJobFromPriorityToActive(KEYS[3], KEYS[2], KEYS[10])
             if jobId then
                 return prepareJobForProcessing(prefix, KEYS[6], eventStreamKey, jobId,
-                                               timestamp, maxJobs,
+                                               timestamp, maxJobs, markerKey,
                                                opts)
             end
         end